+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+* Postmaster and Engrams should be sorted exactly like in game now.
+
 ## 6.84.0 <span class="changelog-date">(2021-09-26)</span>
 
 * Items in the Compare view no longer move around according to the character they're on.